Use pycdlib to read all the needed data without the need for mounting
There's various reasons why this is needed, but the main one being speed and remove the need for auto-mounting if the device wasn't mounted.

With this method of reading the disc's data, you never need to mount the device, we read all the needed data right from the device's sectors.

I tested current master branch and this commit and confirmed that this is producing 1:1 identical checksums.

I updated the setup.py to include the pycdlib requirement as well as slightly updated README